+++ spec/lib/glimmer/dsl/swt/glimmer__table__spec.rb
@@ -26,6 +26,10 @@ module GlimmerSpec
       class Person
         attr_accessor :name, :age, :adult, :dob, :salary
         
+        def formatted_dob
+          dob.strftime("%Y-%m-%d %I:%M:%S %p")
+        end
+        
         def name_options
           ["Ava Fang", "Bruce Ting", "Julia Fang"]
         end
@@ -550,6 +554,88 @@ module GlimmerSpec
           selection = @table.swt_widget.getSelection
           expect(selection.size).to eq(1)
           expect(selection.first.getData).to eq(person2)
+        end
+        
+        it "triggers configured column-specific table widget date_time editing on specified table item" do
+          @target = shell {
+            @table = table(:editable) {
+              table_column {
+                text "DOB"
+                width 120
+                editor :date_time, :date, :time, property: :dob
+              }
+              items bind(group, :people), column_properties(:formatted_dob)
+              selection bind(group, :selected_person)
+            }
+          }
+          
+          expect(@table.table_editor_widget_proxy).to be_nil
+          @write_done = false
+          @table.edit_table_item(
+            @table.swt_widget.items[1],
+            0,
+            before_write: lambda {
+              expect(@table.edit_in_progress?).to eq(true)
+            },
+            after_write: lambda { |edited_table_item|
+              expect(edited_table_item.getText(0)).to eq('1978-11-27 10:47:25 PM')
+              @write_done = true
+            }
+          )
+          expect(@table.table_editor_widget_proxy).to_not be_nil
+          expect(@table.table_editor_widget_proxy.swt_widget).to be_a(DateTime)
+          expect(@table.table_editor_widget_proxy.date_time).to eq(person2.dob)
+          @table.table_editor_widget_proxy.date_time = Time.new(1968, 10, 26, 21, 46, 24)
+
+          event = Event.new
+          event.keyCode = Glimmer::SWT::SWTProxy[:cr]
+          event.doit = true
+          event.character = "\n"
+          event.display = @table.table_editor_widget_proxy.swt_widget.getDisplay
+          event.item = @table.table_editor_widget_proxy.swt_widget
+          event.widget = @table.table_editor_widget_proxy.swt_widget
+          event.type = Glimmer::SWT::SWTProxy[:keydown]
+          @table.table_editor_widget_proxy.swt_widget.notifyListeners(Glimmer::SWT::SWTProxy[:keydown], event)
+
+          expect(@write_done).to eq(true)
+          expect(@table.edit_in_progress?).to eq(false)
+          expect(@cancel_done).to be_nil
+          expect(person2.dob).to eq(Time.new(1968, 10, 26, 21, 46, 24))
+                          
+#           expect(@table.table_editor_widget_proxy).to be_nil
+#           @write_done = false
+#           @table.edit_table_item(
+#             @table.swt_widget.items[1],
+#             0,
+#             before_write: lambda {
+#               expect(@table.edit_in_progress?).to eq(true)
+#             },
+#             after_write: lambda { |edited_table_item|
+#               expect(edited_table_item.getText(0)).to eq('1')
+#               @write_done = true
+#             }
+#           )
+#           expect(@table.table_editor_widget_proxy).to_not be_nil
+#           expect(@table.table_editor_widget_proxy.swt_widget).to be_a(Spinner)
+#
+#           @table.table_editor_widget_proxy.swt_widget.setSelection(1)
+#           event = Event.new
+#           event.doit = true
+#           event.character = "\n"
+#           event.display = @table.table_editor_widget_proxy.swt_widget.getDisplay
+#           event.item = @table.table_editor_widget_proxy.swt_widget
+#           event.widget = @table.table_editor_widget_proxy.swt_widget
+#           event.type = Glimmer::SWT::SWTProxy[:selection]
+#           @table.table_editor_widget_proxy.swt_widget.notifyListeners(Glimmer::SWT::SWTProxy[:selection], event)
+#           expect(@write_done).to eq(true)
+#           expect(@table.edit_in_progress?).to eq(false)
+#           expect(@cancel_done).to be_nil
+#           expect(person2.age).to eq(1)
+#
+          ## test that it maintains selection
+#           selection = @table.swt_widget.getSelection
+#           expect(selection.size).to eq(1)
+#           expect(selection.first.getData).to eq(person2)
         end
         
         it "triggers configured column-specific table widget combo editing on specified table item" do